Triangle follows finger

+++ app/src/main/cpp/android_main.cpp
@@ -77,8 +77,8 @@ static int32_t engine_handle_input(struct android_app *app, AInputEvent *event) 
     struct engine *engine = (struct engine *) app->userData;
     if (AInputEvent_getType(event) == AINPUT_EVENT_TYPE_MOTION) {
         engine->animating = 1;
-        engine->state.x = AMotionEvent_getX(event, 0);
-        engine->state.y = AMotionEvent_getY(event, 0);
+        engine->state.x = (int32_t)AMotionEvent_getX(event, 0);
+        engine->state.y = (int32_t)AMotionEvent_getY(event, 0);
         LOGI("input: x=%d y=%d", engine->state.x, engine->state.y);
         return 1;
     }
